Rename "app" folder to "mchub"
