+++ .dockerignore
@@ -1,5 +1,5 @@
 frontend/node_modules
-app/.pytest_cache
+mchub/.pytest_cache
 .vscode
 
 docs
