+++ .vscode/settings.json
@@ -1,7 +1,7 @@
 {
     "python.pythonPath": "/usr/local/bin/python",
     "python.testing.pytestArgs": [
-        "app/tests"
+        "tests"
     ],
     "python.testing.unittestEnabled": false,
     "python.testing.nosetestsEnabled": false,
