+++ mchub/models/cloud/cloud_manager.py
