+++ mchub/models/cloud/openstack_manager.py
