+++ mchub/models/constants.py
@@ -16,7 +16,7 @@ TERRAFORM_REQUIRED_VERSION = ">= 1.1.0"
 DEFAULT_CLOUD = environ.get("OS_CLOUD", "openstack")
 RUN_PATH = environ.get("MCH_RUN_PATH", getcwd())
 CLUSTERS_PATH = environ.get("MCH_CLUSTERS_PATH", path.join(RUN_PATH, "clusters"))
-APP_PATH = environ.get("MCH_APP_PATH", path.join(RUN_PATH, "app"))
+APP_PATH = environ.get("MCH_APP_PATH", path.join(RUN_PATH, "mchub"))
 DIST_PATH = environ.get("MCH_DIST_PATH", path.join(RUN_PATH, "dist"))
 DATABASE_PATH = environ.get("MCH_DATABASE_PATH", path.join(RUN_PATH, "database"))
 SCHEMA_MIGRATIONS_DIRECTORY = path.join(APP_PATH, "database", "migrations")
