+++ mchub/models/magic_castle/magic_castle_configuration.py
