+++ mchub/models/puppet/provisioning_manager.py
