+++ tests/test_helpers.py
@@ -7,10 +7,10 @@ from shutil import rmtree, copytree
 from typing import Callable
 
 
-from app.database.schema_manager import SchemaManager
-from app.database.database_manager import DatabaseManager
-from app.models.user.authenticated_user import AuthenticatedUser
-from app.models.constants import DEFAULT_CLOUD
+from mchub.database.schema_manager import SchemaManager
+from mchub.database.database_manager import DatabaseManager
+from mchub.models.user.authenticated_user import AuthenticatedUser
+from mchub.models.constants import DEFAULT_CLOUD
 
 from . mocks.openstack.openstack_connection_mock import OpenStackConnectionMock
 
@@ -34,7 +34,7 @@ def teardown_mock_clusters(cluster_names):
 @pytest.fixture(autouse=True)
 def database_connection(mocker):
     # Using an in-memory database for faster unit tests with less disk IO
-    mocker.patch("app.database.database_manager.DATABASE_FILE_PATH", new=":memory:")
+    mocker.patch("mchub.database.database_manager.DATABASE_FILE_PATH", new=":memory:")
 
     with DatabaseManager.connect() as database_connection:
         # The database :memory: only exist within a single connection.
@@ -50,7 +50,7 @@ def database_connection(mocker):
                 pass
 
         mocker.patch(
-            "app.database.database_manager.DatabaseManager.connect",
+            "mchub.database.database_manager.DatabaseManager.connect",
             return_value=MockDatabaseConnection(),
         )
 
@@ -176,9 +176,9 @@ def admin() -> Callable[[sqlite3.Connection], AuthenticatedUser]:
 
 @pytest.fixture(autouse=True)
 def mock_clusters_path(mocker):
-    mocker.patch("app.models.constants.CLUSTERS_PATH", new=MOCK_CLUSTERS_PATH)
+    mocker.patch("mchub.models.constants.CLUSTERS_PATH", new=MOCK_CLUSTERS_PATH)
     mocker.patch(
-        "app.models.magic_castle.magic_castle.CLUSTERS_PATH",
+        "mchub.models.magic_castle.magic_castle.CLUSTERS_PATH",
         new=MOCK_CLUSTERS_PATH,
     )
 
@@ -212,11 +212,11 @@ def disable_provisionning_polling(mocker):
     To avoid this behaviour, we mock ProvisioningManager.is_busy.
     """
     mocker.patch(
-        "app.models.puppet.provisioning_manager.ProvisioningManager.is_busy",
+        "mchub.models.puppet.provisioning_manager.ProvisioningManager.is_busy",
         return_value=True,
     )
 
 
 @pytest.fixture
 def fake_successful_subprocess_run(mocker):
-    mocker.patch("app.models.magic_castle.magic_castle.run", return_value=None)
+    mocker.patch("mchub.models.magic_castle.magic_castle.run", return_value=None)
