+++ tests/unit/user/test_authentified_user.py
@@ -1,8 +1,8 @@
 import pytest
 
-from app.models.constants import DEFAULT_CLOUD
-from app.models.magic_castle.cluster_status_code import ClusterStatusCode
-from app.exceptions.invalid_usage_exception import ClusterNotFoundException
+from mchub.models.constants import DEFAULT_CLOUD
+from mchub.models.magic_castle.cluster_status_code import ClusterStatusCode
+from mchub.exceptions.invalid_usage_exception import ClusterNotFoundException
 
 from ... test_helpers import *  # noqa
 from ... mocks.configuration.config_mock import config_auth_saml_mock  # noqa;
